Attempt to debounce mousemove
The debouncer package from Gizra doesn't work if you set the throttle
threshold under 300 ms or so — you can't see the position of the node
you're placing until you click. I'm not sure why this is — perhaps this
is an "only in development" sort of thing — but there doesn't seem to be
any better options for debouncing at the moment.