Use SVG API to draw editor cursor
This took a little while to figure out. First, I had to figure out how
to use the SVG API, since Elm doesn't expose access to this. So, to
address this, we actually access SVG within JS-land, and we then set up
a communication channel between Elm-land and JS-land. But we can't just
use ports, because ports are one-way. So, we use a port to talk to JS,
and here we can calculate where the cursor is supposed to be based on
the contents of the SVG text element. Then we set up a custom event
listener to talk to Elm again, so that we can send back that cursor
position and draw the cursor on the screen.