Extract NodeContent update/view code to NodeContent
This was particularly tricky, because you have to make sure all of the
types line up, particularly in `update` and `view`. Messages that you
generate in child components have to be wrapped in the parent Msg, and
messages generated in parent components have to be wrapped in the child
Msgs. `Svg.map`, `Cmd.map` and `Sub.map` can help with this.

Resources that helped me:

* https://stackoverflow.com/a/38222804
* https://github.com/halfzebra/elm-examples/tree/master/examples/fractal-architecture
* https://github.com/sporto/elm-example-app
* https://github.com/rtfeldman/elm-spa-example