Prevent Alt-Left and -Right from refreshing page
Basically, we needed to attach a `preventDefault` to the global
`onKeyDown`. However, the `onKeyDown` function from Browser.Events
doesn't let us do this. The VirtualDom module does have an `on` function
that may be useful for this but unfortunately it doesn't expose this. So
we have to resort to calling `preventDefault` via JS-land.

BTW, I tried creating an `effect` module to mimic what VirtualDom was
doing, but since 0.19, only official packages can do that:
<https://discourse.elm-lang.org/t/native-code-in-0-19/826>