Add Cypress tests
Also add an "init" event to node content which will fire when the node
is added to the page, at which point we will start listening to key
events. The Cypresss tests run through everything really fast and so
without this procedure there was a chance that the key event listener
was added before the node was registered internally.