Don't import CSS via JS
Remember that we have to generate Elm style files from CSS module files
in order to be able to import those styles from within Elm. I couldn't
figure out exactly why, but this process seems to work better when
running `parcel build` vs. `parcel serve` if we import the CSS files in
HTML and CSS rather than in JS. This is necessary in order to be able to
run the Cypress tests on CI.