Remove Elm compiled CSS modules in favor of BEM
The Parcel plugin that is supposed to compile Elm files from CSS files
doesn't work very well. It works out that when Parcel reads files and
builds the asset graph, it will do so asychronously. That means it tries
to build the CSS files at the same time it tries to build the Elm file.
That's not going to work if the Elm CSS module files haven't been
compiled yet, because Elm will complain that it's trying to find files
that aren't there. We can get this to work locally, but on CI this
always happens because we start from a clean slate.

So, instead of trying to wrangle this and get this to work, just go with
BEM. We don't have that much CSS anyway.